Adding first attempt to solve JP-2297 by modifying the ramp_fitting.utils.dq_comopress_final function.
Adding a test for the new dq_compress_final function.

Changing a line to make it clearer a uint32 is required for the computation.

Changing style failures in ramp fitting test file.

Adding to the change log.

Removing a confusing comment and expanding the comment explaining the test to make it more understandable.

Correcting style fails.