Update to handle single good group
The baseline code assumes that if there is only one group in an integration it must be the first group. This is not true when MIRI is flagging more than one frame like happens in the RSCD step.
Also, there was the assumption that only saturation could create the single good group but with CR neighbor flagging turned on you can have the first group good and the second group have a jump. So, jumps need to be added to the total number of groups that are bad.