remove warning filter
Tests were unable to produce a situation
where the filtered warning could be produced.
The goal was to add a more specific filter
(one that uses a message). If the warning
can be reproduced the filter with a message
can be re-added. It's possible this warning
is no longer produced and so this commit
removes the filter.